Created an interface for controllers for mediaplayer apps

+++ SharpCaster/Models/ChromecastRequests/VolumeDataObject.cs
@@ -7,7 +7,7 @@ namespace SharpCaster.Models.ChromecastRequests
     public class VolumeDataObject
     {
         [DataMember(Name = "level")]
-        public float? Level { get; set; }
+        public double? Level { get; set; }
 
         [DataMember(Name = "muted")]
         public bool? Muted { get; set; }
